feat: implementar tratamento de erro na entrada de notas e melhorias no cálculo da média
- Adicionado tratamento de erro com `double.TryParse()` para evitar exceções ao converter entradas inválidas.
- Removido o uso desnecessário de `double?`, já que `double.Parse()` sempre retorna um valor válido (ou lança exceção).
- Implementado um loop `while` para garantir que o usuário insira valores numéricos válidos antes de prosseguir.
- Melhorada a legibilidade do código, removendo verificações impossíveis (`if (nota == null)`).
- Ajustada a formatação da média final (`mediaTotal:F1`) para exibição padronizada com uma casa decimal.

+++ CSharp/EstruturaSequencial/Sequencial.cs
@@ -303,7 +303,7 @@ namespace CSharp.EstruturaSequencial
             nome2 = vetorUser2[0];
             idade2 = int.Parse(vetorUser2[1]);
 
-            media = (idade1 + idade2) / 2.0;
+            media = (double) (idade1 + idade2) / 2.0;
 
             Console.WriteLine($"A idade média de {nome1} e {nome2} é de {media.ToString("F1")} anos");
